Upgraded Makefiles to .NET 2.0.
git-svn-id: svn://svn.code.sf.net/p/opendicom/code@130 01fde7d9-2d1c-0410-b6eb-97e1bc93dc97

+++ opendicom-sharp/src/openDicom/Encoding/OF.cs
@@ -44,7 +44,7 @@ namespace openDicom.Encoding
 
         protected override Array DecodeImproper(byte[] bytes)
         {
-            float[] floatValue = new float[(int) Math.Floor(bytes.Length / 4)];
+            float[] floatValue = new float[(int) Math.Floor((double) bytes.Length / 4)];
             byte[] buffer = new byte[4];
             for (int i = 0; i < floatValue.Length; i++)
             {
